Rename JAVA package
