+++ example/android/app/src/main/java/com/baseflow/flutter/plugin/geolocator/example/MainActivity.java
@@ -1,4 +1,4 @@
-package com.baseflow.fluttergeolocatorexample;
+package com.baseflow.flutter.plugin.geolocator.example;
 
 import android.os.Bundle;
